sql,sqlccl: move partitioning creation code to sqlccl
Move partitioning creation code to sqlccl so that we have a logical
place to insert an enterprise license check. (Though CREATE TABLE ...
PARTITION BY ... presently succeeds in an OSS binary, the created table
is of essentially no use, as setting zone configs on partitions requires
a CCL binary.)

Release note: None
