+++ pkg/ccl/sqlccl/logictestccl/testdata/logic_test/partitioning_index
