+++ pkg/sql/create_index.go
@@ -71,7 +71,7 @@ func (n *createIndexNode) startExec(params runParams) error {
 		return err
 	}
 	if n.n.PartitionBy != nil {
-		partitioning, err := createPartitionedBy(params.ctx, params.p.ExecCfg().Settings,
+		partitioning, err := CreatePartitioning(params.ctx, params.p.ExecCfg().Settings,
 			params.evalCtx, n.tableDesc, &indexDesc, n.n.PartitionBy)
 		if err != nil {
 			return err
